Create DialogBase class
And use it in LoginView class and RegisterMemberView class
