+++ src/javaiina/LoginView.java
@@ -22,7 +22,7 @@ import javax.swing.JTextField;
 import javax.swing.border.Border;
 import javax.swing.border.EmptyBorder;
 
-public class LoginView extends JDialog
+public class LoginView extends DialogBase
 {
     private static final long serialVersionUID = -7678380755203024447L;
     
@@ -30,19 +30,6 @@ public class LoginView extends JDialog
     private static final int DefaultWindowHeight = 320;
     
     private static final String DefaultWindowTitle = "Login";
-    
-    private DialogResult mResult;
-    
-    private JPanel mPanelHeader;
-    private JPanel mPanelCenter;
-    private JPanel mPanelBottom;
-    
-    private Font mFontHeaderTitle;
-    private JLabel mLabelHeaderTitle;
-    private Font mFontHeaderDescription;
-    private JLabel mLabelHeaderDescription;
-    
-    private Font mFontLabel;
     
     private JLabel mLabelUserId;
     private JTextField mTextBoxUserId;
@@ -58,59 +45,26 @@ public class LoginView extends JDialog
     {
         super(parentFrame);
         
-        this.initializeComponent();
-        this.addEventHandler();
-        
         this.setMinimumSize(new Dimension(
             LoginView.DefaultWindowWidth, LoginView.DefaultWindowHeight));
         this.setTitle(LoginView.DefaultWindowTitle);
         this.setLocationRelativeTo(this.getParent());
     }
     
-    private void initializeComponent()
+    @Override
+    protected void initializeComponent()
     {
-        /* Border */
-        Border defaultBorder = new EmptyBorder(5, 5, 5, 5);
+        super.initializeComponent();
         
         /* Insets */
         Insets defaultInsets = new Insets(5, 5, 5, 5);
         Insets emptyInsets = new Insets(0, 0, 0, 0);
         
-        /* Container */
-        this.mPanelHeader = new JPanel();
-        this.mPanelHeader.setLayout(new BoxLayout(this.mPanelHeader, BoxLayout.Y_AXIS));
-        this.mPanelHeader.setBorder(defaultBorder);
-        this.getContentPane().add(this.mPanelHeader, BorderLayout.NORTH);
-        
-        this.mPanelCenter = new JPanel();
-        this.mPanelCenter.setLayout(new GridBagLayout());
-        this.getContentPane().add(this.mPanelCenter, BorderLayout.CENTER);
-        
-        this.mPanelBottom = new JPanel();
-        this.mPanelBottom.setLayout(new GridBagLayout());
-        this.getContentPane().add(this.mPanelBottom, BorderLayout.SOUTH);
-        
         /* Header */
-        this.mFontHeaderTitle = new Font(Font.DIALOG, Font.BOLD, 14);
-        this.mLabelHeaderTitle = new JLabel();
-        this.mLabelHeaderTitle.setText("Login");
-        this.mLabelHeaderTitle.setFont(this.mFontHeaderTitle);
-        this.mPanelHeader.add(this.mLabelHeaderTitle);
-        
-        this.mPanelHeader.add(Box.createVerticalStrut(5));
-        
-        this.mFontHeaderDescription = new Font(Font.DIALOG, Font.PLAIN, 12);
-        this.mLabelHeaderDescription = new JLabel();
-        this.mLabelHeaderDescription.setText(
-            "Login to Javaiina - Costume rental management system.");
-        this.mLabelHeaderDescription.setFont(this.mFontHeaderDescription);
-        this.mPanelHeader.add(this.mLabelHeaderDescription);
-        
-        this.mPanelHeader.add(Box.createVerticalStrut(15));
+        this.setHeaderTitle("Login");
+        this.setHeaderDescription("Login to Javaiina - Costume rental management system.");
         
         /* Center */
-        this.mFontLabel = new Font(Font.DIALOG, Font.PLAIN, 12);
-        
         GridBagConstraints layoutConstraints = new GridBagConstraints();
         layoutConstraints.anchor = GridBagConstraints.WEST;
         
@@ -169,24 +123,13 @@ public class LoginView extends JDialog
         this.mPanelBottom.add(this.mButtonCancel, layoutConstraints);
     }
     
-    private void addEventHandler()
+    @Override
+    protected void addEventHandler()
     {
+        super.addEventHandler();
+        
         this.mButtonLogin.addActionListener(e -> this.onButtonLoginClick());
         this.mButtonCancel.addActionListener(e -> this.onButtonCancelClick());
-    }
-    
-    private GridBagConstraints setLayoutConstraints(
-        GridBagConstraints layoutConstraints,
-        int gridx, int gridy, double weightx, double weighty, Insets insets, int fill)
-    {
-        layoutConstraints.gridx = gridx;
-        layoutConstraints.gridy = gridy;
-        layoutConstraints.weightx = weightx;
-        layoutConstraints.weighty = weighty;
-        layoutConstraints.insets = insets;
-        layoutConstraints.fill = fill;
-        
-        return layoutConstraints;
     }
     
     private void onButtonLoginClick()
@@ -212,11 +155,6 @@ public class LoginView extends JDialog
     public void setInputValidator(InputValidator<LoginView> inputValidator)
     {
         this.mInputValidator = inputValidator;
-    }
-    
-    public DialogResult getResult()
-    {
-        return this.mResult;
     }
     
     public String getUserId()
